Added move.c and move.h files

+++ src/ai.c
@@ -1,20 +1,19 @@
 #include <stdio.h>
-#include <stdint.h>
 #include "board.h"
 
 
-uint8_t probability(uint8_t rows, uint8_t cols, uint16_t gb[rows][cols]) {
+int probability(int rows, int cols, int gb[rows][cols]) {
 	return 0;
 }
 
-void evaluate(uint8_t rows, uint8_t cols, uint16_t gb[rows][cols], uint16_t sb[rows][cols], uint8_t size) {
-	for (uint8_t i = 0; i < rows; i++) {
-		for (uint8_t j = 0; j < cols; j++) {
-			uint16_t right = 1;
-			uint16_t left = 1;
-			uint16_t down = 1;
-			uint16_t up = 1;
-			for (uint16_t s = 0; s < size; s++) {
+void evaluate(int rows, int cols, int gb[rows][cols], int sb[rows][cols], int size) {
+	for (int i = 0; i < rows; i++) {
+		for (int j = 0; j < cols; j++) {
+			int right = 1;
+			int left = 1;
+			int down = 1;
+			int up = 1;
+			for (int s = 0; s < size; s++) {
 				// right
 				if (j <= cols-size) {
 					right *= gb[i][j+s];
